+++ packages/x-components/src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './directional-focus-navigation.service'
+export * from './results-enrichment.service'
 export * from './services.types'